+++ sdk/search/search-documents/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Bugs Fixed
 
 - Converted the complex fields correctly within the Search Fields. Please refer [#16489](https://github.com/Azure/azure-sdk-for-js/issues/16489) for more details.
+- Fixed the typos `anayzerName` to `analyzerName` and `normalizerNames` to `normalizerName` in `convertFieldsToPublic` method of `serviceUtils.ts`.
 
 ### Other Changes